requirements.txt
absl-py==0.8.1
astor==0.8.0
cycler==0.10.0
decorator==4.4.1
gast==0.3.2
grpcio==1.25.0
h5py==2.10.0
imageio==2.6.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.14.5
pandas==0.25.3
Pillow==6.2.1
plotly==4.2.1
protobuf==3.10.0
pymia==0.2.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
retrying==1.3.3
scikit-image==0.16.2
scipy==1.3.2
SimpleITK==1.2.3
six==1.13.0
tensorboard==1.10.0
tensorboardX==1.9
tensorflow-gpu==1.10.0
termcolor==1.1.0
torch==1.3.1
Werkzeug==0.16.0